opensll: add cc req and handle shared opt

+++ recipes/openssl/all/conanfile.py
@@ -4,29 +4,36 @@ from build import *
 class OpensslRecipe(Recipe):
     description = "TLS/SSL and crypto library"
     license = "BSD"
+    options = {"shared": [True, False], "bootstrap": [True, False]}
+    default_options = {"shared": True, "bootstrap": False}
     build_requires = (
+        "cc/[^1.0.0]",
         "make/[^4.3]",
         "perl/[^5.30.0]",
     )
-    requires = ("ca-certificates/[^20191127]",)
+
+    def requirements(self):
+        if not self.options.bootstrap:
+            self.requires("ca-certificates/[^20191127]")
 
     def source(self):
         self.get(f"https://github.com/openssl/openssl/archive/openssl-{self.version}.tar.gz")
 
     def build(self):
-        self.run("mv Configure configure", cwd=f"{self.name}-{self.version}")
-
-        args = ["shared", "no-ssl3-method"]
-        if self.settings.arch_build == "x86_64":
+        args = ["no-ssl3-method"]
+        if self.options.shared:
+            args.append("shared")
+        if self.settings.arch == "x86_64":
             args += ["linux-x86_64", "enable-ec_nistp_64_gcc_128"]
-        elif self.settings.arch_build == "armv8":
+        elif self.settings.arch == "armv8":
             args += ["linux-aarch64", "no-afalgeng"]
-        self.autotools(args)
+        self.exe("./Configure", args)
+        self.make()
 
-        # Remove static libs
-        libs = ["crypto", "ssl"]
-        for lib in libs:
-            os.remove(os.path.join(self.package_folder, "lib", f"lib{lib}.a"))
+        if not self.options.shared:
+            libs = ["crypto", "ssl"]
+            for lib in libs:
+                os.remove(os.path.join(self.package_folder, "lib", f"lib{lib}.a"))
 
     def package_info(self):
         self.env_info.SSL_CERT_DIR = os.path.join(self.deps_cpp_info["ca-certificates"].rootpath, "etc", "ssl", "certs")
